project-sia/requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0 